final css style rules
The rule is for the navigation and has an id of bottom and the rule is making the the bottom nav a flex-box and doesn't have wrap.  The rule is for the unordered list thats in the navigation with the id of bottom and the rule is making the growth rate shrink rate both 1 and the basis 150 pixels. The rule is foe the list item background thats nested in the unordered list that is in the navigation with id of top and the rule is making the background an image. The rule is for the anchors thats nested in the paragraph tag that has id of links and the rule is making the text white, the background color is red, the border radius 20 pixels, and no text-decoration. The rule is fo the unordered list that is nested in the nav with the id of top and the rule is making the nav flex-box with background color, a height of 50 pixels, and makes the content horizontal and vertical centered. Ready for Review